% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{HDR data until 2020}
\format{
A data frame 6 variables:
\describe{
  \item{indicator_id}{Indicator ID}
  \item{indicator_name}{Indicator name}
  \item{country_code}{country code}
  \item{country_name}{country name}
  \item{year}{}
  \item{value}{Indicator value}
}
}
\usage{
df
}
\description{
HDR data until 2020
}
\keyword{datasets}
