% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdr_indicators}
\alias{hdr_indicators}
\title{HDR Indicators}
\format{
A data frame 10 variables:
\describe{
  \item{ID}{ID Indicator}
  \item{Indicator Name}{Indicator Name}
}

A data frame 10 variables:
\describe{
  \item{ID}{ID Indicator}
  \item{Indicator Name}{Indicator Name}
}
}
\usage{
hdr_indicators

hdr_indicators
}
\description{
HDR Indicators

HDR data until 2020
}
\keyword{datasets}
