% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twd.R
\name{twd}
\alias{twd}
\title{requires xts, lubridate, dplyr
read more about total water deficit 10.1029/WR016i002p00297}
\usage{
twd(df.hist, df.x)
}
\arguments{
\item{df.hist}{dataframe of historical monthly reservoir fill volume in % column name is "rs". First column must be of class POSIXct}

\item{df.x}{dataframe with same structure of df.hist with values for the months which we are calculating the index. Should have more than one row, ie more than one month}
}
\description{
requires xts, lubridate, dplyr
read more about total water deficit 10.1029/WR016i002p00297
}
