% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hdi_data.R
\name{update_hdi_data}
\alias{update_hdi_data}
\title{Extract Human Development Index Data}
\usage{
update_hdi_data()
}
\description{
This function extract the  small dataset each year which consists of approximately 144 indicators by country and year.
The dowloaded data contains 6 variables : indicator_id, indicator_name, country_code, country_name, year, value
}
