\name{hdi-package}
\alias{hdi-package}
\docType{package}
\title{hdi}
\description{
Perform inference in high-dimensional (generalized) linear models using various approaches.
}
\details{
\tabular{ll}{
Package: \tab hdi\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-02\cr
License: \tab GPL \cr
}
This is a very early test release!
}
\author{
Lukas Meier

Maintainer: Lukas Meier <meier@stat.math.ethz.ch>
}
\references{
  Meinshausen, N., Meier, L. and B\"uhlmann, P. (2009),
  \emph{P-values for high-dimensional regression}, Journal of the
  American Statistical Association 104, 1671-1681.

  Meinshausen, N. and B\"uhlmann, P. (2010), \emph{Stability selection
  (with discussion)}, Journal of the Royal Statistical Society: Series B,
  72, 417-473.}
\keyword{ package }

