\name{hdi-package}
\alias{hdi-package}
\docType{package}
\title{hdi}
\description{
Perform inference in high-dimensional (generalized) linear models using
various approaches.
}
\details{
\tabular{ll}{
Package: \tab hdi\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-02\cr
License: \tab GPL \cr}
}
\author{
  Lukas Meier, 
  Ruben Dezeure, 
  Nicolai Meinshausen, 
  Peter B\"uhlmann, 
  Maintainer: Lukas Meier <meier@stat.math.ethz.ch>
}
\references{
  Dezeure, R., B\"uhlmann, P., Meier, L. and Meinshausen,
  N. (2014). \emph{High-dimensional inference: confidence intervals, p-values
  and R-software hdi}, Statistical Science. To appear.
  
  Meinshausen, N., Meier, L. and B\"uhlmann, P. (2009),
  \emph{P-values for high-dimensional regression}, Journal of the
  American Statistical Association 104, 1671-1681.

  Meinshausen, N. (2013), \emph{Group-bound: confidence intervals for
  groups of variables in sparse high-dimensional regression without
  assumptions on the design}. Journal of the Royal Statistical Society,
  Series B. To appear.

  Meinshausen, N. and B\"uhlmann, P. (2010), \emph{Stability selection
  (with discussion)}, Journal of the Royal Statistical Society: Series B,
  72, 417-473.}
\keyword{ package }

